+++ example/main.dart
@@ -38,6 +38,7 @@ class _ExampleLauncherPageState extends State<ExampleLauncherPage> {
   String signUpResultText = '<Not yet signed up>';
   String profilePageResultText = '';
   String paymentMethodsPageResultText = '';
+  String driversLicenseResultText = '';
 
   ///
   /// Having the [MLogin] as an object with longer lifecycle is recommended but
@@ -50,6 +51,7 @@ class _ExampleLauncherPageState extends State<ExampleLauncherPage> {
     clientId: 'm-login-demo-app-k',
     callbackUrlScheme: 'k.de.swm.login.app',
     redirectUri: 'k.de.swm.login.app:/oauth2redirect/example',
+    idVerificationRedirectUri: 'k.de.swm.login.app:/oauth2redirect/example',
   );
 
   @override
@@ -79,6 +81,11 @@ class _ExampleLauncherPageState extends State<ExampleLauncherPage> {
                 child: const Text('Profile Page'),
               ),
               Text(profilePageResultText),
+              ElevatedButton(
+                onPressed: _openDriversLicenseVerificationPage,
+                child: const Text('Drivers License Verification'),
+              ),
+              Text(driversLicenseResultText),
               Container(
                 color: Colors.grey,
                 padding: const EdgeInsets.all(10),
@@ -152,6 +159,17 @@ class _ExampleLauncherPageState extends State<ExampleLauncherPage> {
     setState(() {
       paymentMethodsPageResultText =
           'Payment methods page was shown, result: $result';
+    });
+  }
+
+  _openDriversLicenseVerificationPage() async {
+    setState(() {
+      driversLicenseResultText = 'Opening Drivers License page...';
+    });
+
+    final result = await mLogin.openDriverLicenseVerification();
+    setState(() {
+      driversLicenseResultText = 'Profile was shown, result: $result';
     });
   }
 }